Updating 2.1 Queued Hosted Service Docs to use BackgroundService
Updating Queued Hosted Service to use Background Service.

Re-purposed text for description from https://docs.microsoft.com/en-us/dotnet/standard/microservices-architecture/multi-container-microservice-net-applications/background-tasks-with-ihostedservice

Modified the BackgroundService Link to use xref (I'm under the assumption that just linking the Class is enough based on the IHosted Service implementation from other docs)
