+++ aspnetcore/fundamentals/host/hosted-services/samples/2.x/BackgroundTasksSample-WebHost/Services/QueuedHostedService.cs
@@ -7,14 +7,12 @@ using Microsoft.Extensions.Logging;
 namespace BackgroundTasksSample.Services
 {
     #region snippet1
-    public class QueuedHostedService : IHostedService
+    public class QueuedHostedService : BackgroundService
     {
-        private CancellationTokenSource _shutdown = 
-            new CancellationTokenSource();
         private Task _backgroundTask;
         private readonly ILogger _logger;
 
-        public QueuedHostedService(IBackgroundTaskQueue taskQueue, 
+        public QueuedHostedService(IBackgroundTaskQueue taskQueue,
             ILoggerFactory loggerFactory)
         {
             TaskQueue = taskQueue;
@@ -23,42 +21,22 @@ namespace BackgroundTasksSample.Services
 
         public IBackgroundTaskQueue TaskQueue { get; }
 
-        public Task StartAsync(CancellationToken cancellationToken)
+        protected async override Task ExecuteAsync(CancellationToken stoppingToken)
         {
-            _logger.LogInformation("Queued Hosted Service is starting.");
-
-            _backgroundTask = Task.Run(BackgroundProceessing);
-
-            return Task.CompletedTask;
-        }
-
-        private async Task BackgroundProceessing()
-        {
-            while (!_shutdown.IsCancellationRequested)
+            while (!stoppingToken.IsCancellationRequested)
             {
-                var workItem = 
-                    await TaskQueue.DequeueAsync(_shutdown.Token);
-
+                var workItem =
+                    await TaskQueue.DequeueAsync(stoppingToken);
                 try
                 {
-                    await workItem(_shutdown.Token);
+                    await workItem(stoppingToken);
                 }
                 catch (Exception ex)
                 {
-                    _logger.LogError(ex, 
+                    this._logger.LogError(ex,
                         $"Error occurred executing {nameof(workItem)}.");
                 }
             }
-        }
-
-        public Task StopAsync(CancellationToken cancellationToken)
-        {
-            _logger.LogInformation("Queued Hosted Service is stopping.");
-
-            _shutdown.Cancel();
-
-            return Task.WhenAny(_backgroundTask, 
-                Task.Delay(Timeout.Infinite, cancellationToken));
         }
     }
     #endregion
